[BACKLOG-2384] Changes to better support loading and caching of dimension members.
1) Introduces a "batch resolution" phase prior to final expression
resolution which collects child name references and loads them at the
same time.  This avoids the one-at-a-time SQL queries for each child
member reference within an expression.  IdBatchResolver is the
relevant class.

2) To better suppport (1), this change also introduces a new caching
mechanism for ChildByNameConstraints.  If the child members { [Bob],
[Joe], [Bill] } have already been cached, MemberCacheHelper is now
smart enough to retrieve any subset of what has formerly been loaded.
This new functionanlity only works with ChildByNameConstraint checks,
since it depends on lookups by name.

3) To better handle small levels (under 300 members), this change also
adds a check for level cardinality when constructing a constraint in
SqlConstraintFactory.  In general, any level with fewer than 300
members will be loaded with a default constraint to avoid repeat
queries to the dimension table.  There are some exceptions to this,
for example degenerate dimensions are assumed to be too expensive for
full table scans, so a Sql constraint is preferred.  This is governed by
a new property called "LevelPreCacheThreshold", which can be set to
0 to disable the new behavior.

Conflicts:
	src/main/mondrian/olap/Query.java
	src/main/mondrian/rolap/SqlConstraintFactory.java
	testsrc/main/mondrian/rolap/BatchTestCase.java
	testsrc/main/mondrian/rolap/NonEmptyTest.java
	testsrc/main/mondrian/test/Main.java
	testsrc/main/mondrian/test/NativeSetEvaluationTest.java