[MONDRIAN-2403] Severe performance issue with reports containing count distinct measures and compound slicers
This commit includes the following changes to reduce the overhead of
distinct-count measures being evaluated in the context of a compound
slicer:
1)  Compound predicates for the slicer are constructed in the
evaluator.  This allows the slicer predicates to be constructed a
single time in most cases, rather than once per cell request.

2) Similarly, the compound predicate string representation for the slicer
is also constructed in the evaluator.

3) AggregateFunDef updated to avoid unnecessary tuple optimization
when tuplelist is supportable within an IN list.  The tuple
optimization algorithm is expensive and we should avoid it when
unnecessary.

To support 1 and 2, the logic in RolapAggregationManager for
constructing compound predicates has been extracted into a new
CompoundPredicateInfo class.  The logic for constructing predicate
strings (i.e. getCompoundPredicateStringList) was extracted from
AggregationKey and included in CompoundPredicateInfo as well.

Conflicts:
	src/main/mondrian/olap/fun/AggregateFunDef.java
	src/main/mondrian/rolap/RolapEvaluator.java