[MONDRIAN-2402]: Mondrian incorrectly resolves <String> Generate(<Set>, <String>, <String>) to <Set> Generate(<Set>, <Set>, <Symbol>)
The cause of the bug:
In mondrian schema a dimention can contain many hierarchies.
In mondrian schema it's allowed hierarchies without name attribute. In this case the hierarchy has the same name as its dimension.
When expression with a dimention is been parsing and resolving and it needs to get the hierarchy of this dimention, then it's returned null for the case when the dimention contains two hierarchy.
This is incorrect because we ignore the hierarchies with the same name as dimention name.
An this is the cause of the issue because the return type for the member Property function is taken exactly from this member resolved hierarchy level.
Please see:
https://github.com/pentaho/mondrian/blob/master/mondrian/src/main/java/mondrian/olap/fun/PropertiesFunDef.java#L113
https://github.com/pentaho/mondrian/blob/master/mondrian/src/main/java/mondrian/olap/fun/PropertiesFunDef.java#L137-L159

The fix is to return the hierarchy of the dimention with the same name as dimention name, when the dimention contains more then one hierarchy.
Added unit tests.